+++ ui/src/components/ModalInvoke/ModalInvoke.less
@@ -1,5 +1,32 @@
 @import '../../assets/styles/theme';
 
+.modal{
+  :global(.ant-modal-close-x) {
+    margin: 0;
+  }
+
+  :global(.ant-btn-primary) {
+    border-color: #f7ad2b;
+    background: #f7ad2b;
+    color: #595959;
+    box-shadow: none;
+
+    &:hover {
+      border-color: #f7ad2b;
+      background: #f7ad2b;
+      color: #595959;
+    }
+    &:focus {
+      border-color: #f7ad2b;
+      background: #f7ad2b;
+      color: #595959;
+    }
+  }
+}
+.multiline {
+  white-space: pre-wrap;
+}
+
 .overwrite  :global(.ant-modal-header) {
   font-weight: 700;
   border-bottom: none;
