--- ui/src/cxboxComponents/ModalInvoke/ModalInvoke.less
@@ -1,27 +1,0 @@
-.modal{
-    :global(.ant-modal-close-x) {
-        margin: 0;
-    }
-
-    :global(.ant-btn-primary) {
-        border-color: #f7ad2b;
-        background: #f7ad2b;
-        color: #595959;
-        box-shadow: none;
-
-        &:hover {
-            border-color: #f7ad2b;
-            background: #f7ad2b;
-            color: #595959;
-        }
-        &:focus {
-            border-color: #f7ad2b;
-            background: #f7ad2b;
-            color: #595959;
-        }
-    }
-}
-.multiline {
-    white-space: pre-wrap;
-}
-
